first pass of building the location framework
the game is organized into regions, which have various connections to other regions. each of these connections has a set of requirements. a requirement is a set of items or events which the player must have in order to fulfill the requirement. each region also has a set of item locations, which each also have a set of requirements as well as the map and event number (not all locations are complete on this front yet) for that location. ideally, the solver will traverse the regions and locations like a graph in order to ensure the game can be completed.