Add speedrunner muscle memory-friendly delays When pressing down to pick up Curly (in the Waterway and in Hell), the natural instinct is to immediately press jump and shoot (in some order) in order to mash through the message box that normally pops up in vanilla. Since those message boxex were removed for rando, this results in the player jumping after picking up Curly (which, if speedrunning Hell, would lose precious time on the clock). Putting in a small delay is intended to eat the jump input, allowing the player to continue onwards as normal.
I also put in a tiny delay at the start of Hell after the message
box, because I found myself often doing an unintentional extra boost
at the start. This doesn't completely solve the issue for me, but
it's slightly better than nothing, so I'll leave this in for now and
come back to it later if I ever find a better solution.