Make Super Missiles strictly an upgrade to the regular Missile Launcher
Turns out that having only 5 Super Missiles as your only weapon is a pain.
Also, it causes problems in cases when there's a required progression item in
the upper Sand Zone Polish spot and no other accessible block-breaker weapon
that you can use to get to it.
(It's still technically possible to reach that item location in such a
scenario, but it requires grinding the missiles up to Level 3, as well as
some luck and a couple of precise jumps. In other words, it sucks.)

The alternative is starting with 10 Super Missiles instead of 5 if you get
the Super Missile chest first, but that would give you an extra 5 missiles,
letting you have 59 at the end instead of 54, which feels wrong.